+++ analyzer/src/main/java/nl/basjes/parse/useragent/AbstractUserAgentAnalyzer.java
@@ -37,6 +37,17 @@ public class AbstractUserAgentAnalyzer extends AbstractUserAgentAnalyzerDirect i
         initializeCache();
     }
 
+    /**
+     * In some cases it was found that a memory leak could occur.
+     * This will wipe all internal data structures to avoid these problems.
+     * After calling this method this instance becomes unusable and cannot be 'repaired'
+     */
+    public synchronized void destroy() {
+        super.destroy();
+        parseCache.values().forEach(UserAgent::destroy);
+        parseCache.clear();
+    }
+
     public static class KryoSerializer extends AbstractUserAgentAnalyzerDirect.KryoSerializer {
         public KryoSerializer(Kryo kryo, Class<?> type) {
             super(kryo, type);